+++ drivers/pinctrl/pfc_rcar.c
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2021 IoT.bzh
+ * Copyright (c) 2021-2023 IoT.bzh
  *
  * SPDX-License-Identifier: Apache-2.0
  *
@@ -10,13 +10,24 @@
 #include <zephyr/arch/cpu.h>
 #include <zephyr/devicetree.h>
 #include <zephyr/drivers/pinctrl.h>
+#include <rcar_pfc_defs.h>
 #include <soc.h>
 #include <zephyr/sys/util.h>
 
-#define PFC_REG_BASE  DT_INST_REG_ADDR(0)
 #define PFC_RCAR_PMMR 0x0
-#define PFC_RCAR_GPSR 0x100
-#define PFC_RCAR_IPSR 0x200
+
+/* Gen3 only has one base address, Gen4 has one per GPIO controller */
+#if defined(CONFIG_SOC_SERIES_RCAR_GEN3)
+static const uint32_t reg_base[] = {DT_INST_REG_ADDR(0)};
+#elif defined(CONFIG_SOC_SERIES_RCAR_GEN4)
+/* swap both arguments */
+#define PFC_REG_ADDRESS(idx, node_id) DT_REG_ADDR_BY_IDX(node_id, idx)
+static const uint32_t reg_base[] = {
+	LISTIFY(DT_NUM_REGS(DT_DRV_INST(0)), PFC_REG_ADDRESS, (,), DT_DRV_INST(0))
+};
+#else
+#error Unsupported SoC Series
+#endif
 
 /*
  * Each drive step is either encoded in 2 or 3 bits.
@@ -30,18 +41,25 @@
 /* Some registers such as IPSR GPSR or DRVCTRL are protected and
  * must be preceded to a write to PMMR with the inverse value.
  */
-static void pfc_rcar_write(uint32_t offs, uint32_t val)
+static void pfc_rcar_write(uint32_t pfc_base, uint32_t offs, uint32_t val)
 {
-	sys_write32(~val, PFC_REG_BASE + PFC_RCAR_PMMR);
-	sys_write32(val, PFC_REG_BASE + offs);
+	sys_write32(~val, pfc_base + PFC_RCAR_PMMR);
+	sys_write32(val, pfc_base + offs);
 }
 
 /* Set the pin either in gpio or peripheral */
-static void pfc_rcar_set_gpsr(uint16_t pin, bool peripheral)
+static void pfc_rcar_set_gpsr(uint32_t pfc_base,
+			      uint16_t pin, bool peripheral)
 {
+#if defined(CONFIG_SOC_SERIES_RCAR_GEN3)
+	/* On Gen3 we have multiple GPSR at one base address */
 	uint8_t bank = pin / 32;
+#elif defined(CONFIG_SOC_SERIES_RCAR_GEN4)
+	/* On Gen4 we have one GPSR at multiple base address */
+	uint8_t bank = 0;
+#endif
 	uint8_t bit = pin % 32;
-	uint32_t val = sys_read32(PFC_REG_BASE + PFC_RCAR_GPSR +
+	uint32_t val = sys_read32(pfc_base + PFC_RCAR_GPSR +
 				  bank * sizeof(uint32_t));
 
 	if (peripheral) {
@@ -49,18 +67,19 @@ static void pfc_rcar_set_gpsr(uint16_t pin, bool peripheral)
 	} else {
 		val &= ~BIT(bit);
 	}
-	pfc_rcar_write(PFC_RCAR_GPSR + bank * sizeof(uint32_t), val);
+	pfc_rcar_write(pfc_base, PFC_RCAR_GPSR + bank * sizeof(uint32_t), val);
 }
 
 /* Set peripheral function */
-static void pfc_rcar_set_ipsr(const struct rcar_pin_func *rcar_func)
+static void pfc_rcar_set_ipsr(uint32_t pfc_base,
+			      const struct rcar_pin_func *rcar_func)
 {
 	uint16_t reg_offs = PFC_RCAR_IPSR + rcar_func->bank * sizeof(uint32_t);
-	uint32_t val = sys_read32(PFC_REG_BASE + reg_offs);
+	uint32_t val = sys_read32(pfc_base + reg_offs);
 
 	val &= ~(0xFU << rcar_func->shift);
 	val |= (rcar_func->func << rcar_func->shift);
-	pfc_rcar_write(reg_offs, val);
+	pfc_rcar_write(pfc_base, reg_offs, val);
 }
 
 static uint32_t pfc_rcar_get_drive_reg(uint16_t pin, uint8_t *offset,
@@ -87,7 +106,8 @@ static uint32_t pfc_rcar_get_drive_reg(uint16_t pin, uint8_t *offset,
  * using DRVCTRLx registers, some pins have 8 steps (3 bits size encoded)
  * some have 4 steps (2 bits size encoded).
  */
-static int pfc_rcar_set_drive_strength(uint16_t pin, uint8_t strength)
+static int pfc_rcar_set_drive_strength(uint32_t pfc_base, uint16_t pin,
+				       uint8_t strength)
 {
 	uint8_t offset, size, step;
 	uint32_t reg, val;
@@ -107,11 +127,11 @@ static int pfc_rcar_set_drive_strength(uint16_t pin, uint8_t strength)
 	 */
 	strength = (strength / step) - 1U;
 	/* clear previous drive strength value */
-	val = sys_read32(PFC_REG_BASE + reg);
+	val = sys_read32(pfc_base + reg);
 	val &= ~GENMASK(offset + size - 1U, offset);
 	val |= strength << offset;
 
-	pfc_rcar_write(reg, val);
+	pfc_rcar_write(pfc_base, reg, val);
 
 	return 0;
 }
@@ -135,7 +155,7 @@ static const struct pfc_bias_reg *pfc_rcar_get_bias_reg(uint16_t pin,
 	return NULL;
 }
 
-int pfc_rcar_set_bias(uint16_t pin, uint16_t flags)
+int pfc_rcar_set_bias(uint32_t pfc_base, uint16_t pin, uint16_t flags)
 {
 	uint32_t val;
 	uint8_t bit;
@@ -146,19 +166,19 @@ int pfc_rcar_set_bias(uint16_t pin, uint16_t flags)
 	}
 
 	/* pull enable/disable*/
-	val = sys_read32(PFC_REG_BASE + bias_reg->puen);
+	val = sys_read32(pfc_base + bias_reg->puen);
 	if ((flags & RCAR_PIN_FLAGS_PUEN) == 0U) {
-		sys_write32(val & ~BIT(bit), PFC_REG_BASE + bias_reg->puen);
+		sys_write32(val & ~BIT(bit), pfc_base + bias_reg->puen);
 		return 0;
 	}
-	sys_write32(val | BIT(bit), PFC_REG_BASE + bias_reg->puen);
+	sys_write32(val | BIT(bit), pfc_base + bias_reg->puen);
 
 	/* pull - up/down */
-	val = sys_read32(PFC_REG_BASE + bias_reg->pud);
+	val = sys_read32(pfc_base + bias_reg->pud);
 	if (flags & RCAR_PIN_FLAGS_PUD) {
-		sys_write32(val | BIT(bit), PFC_REG_BASE + bias_reg->pud);
+		sys_write32(val | BIT(bit), pfc_base + bias_reg->pud);
 	} else {
-		sys_write32(val & ~BIT(bit), PFC_REG_BASE + bias_reg->pud);
+		sys_write32(val & ~BIT(bit), pfc_base + bias_reg->pud);
 	}
 	return 0;
 }
@@ -166,10 +186,23 @@ int pfc_rcar_set_bias(uint16_t pin, uint16_t flags)
 int pinctrl_configure_pin(const pinctrl_soc_pin_t *pin)
 {
 	int ret = 0;
+	uint8_t reg_index;
+	uint32_t pfc_base;
+
+	ret = pfc_rcar_get_reg_index(pin->pin, &reg_index);
+	if (ret) {
+		return ret;
+	}
+
+	if (reg_index >= ARRAY_SIZE(reg_base)) {
+		return -EINVAL;
+	}
+
+	pfc_base = reg_base[reg_index];
 
 	/* Set pin as GPIO if capable */
 	if (RCAR_IS_GP_PIN(pin->pin)) {
-		pfc_rcar_set_gpsr(pin->pin, false);
+		pfc_rcar_set_gpsr(pfc_base, pin->pin, false);
 	} else if ((pin->flags & RCAR_PIN_FLAGS_FUNC_SET) == 0U) {
 		/* A function must be set for non GPIO capable pin */
 		return -EINVAL;
@@ -177,14 +210,14 @@ int pinctrl_configure_pin(const pinctrl_soc_pin_t *pin)
 
 	/* Select function for pin */
 	if ((pin->flags & RCAR_PIN_FLAGS_FUNC_SET) != 0U) {
-		pfc_rcar_set_ipsr(&pin->func);
+		pfc_rcar_set_ipsr(pfc_base, &pin->func);
 
 		if (RCAR_IS_GP_PIN(pin->pin)) {
-			pfc_rcar_set_gpsr(pin->pin, true);
+			pfc_rcar_set_gpsr(pfc_base, pin->pin, true);
 		}
 
 		if ((pin->flags & RCAR_PIN_FLAGS_PULL_SET) != 0U) {
-			ret = pfc_rcar_set_bias(pin->pin, pin->flags);
+			ret = pfc_rcar_set_bias(pfc_base, pin->pin, pin->flags);
 			if (ret < 0) {
 				return ret;
 			}
@@ -192,7 +225,7 @@ int pinctrl_configure_pin(const pinctrl_soc_pin_t *pin)
 	}
 
 	if (pin->drive_strength != 0U) {
-		ret = pfc_rcar_set_drive_strength(pin->pin,
+		ret = pfc_rcar_set_drive_strength(pfc_base, pin->pin,
 						  pin->drive_strength);
 	}
