+++ soc/arm/renesas_rcar/CMakeLists.txt
@@ -1,3 +1,4 @@
 # SPDX-License-Identifier: Apache-2.0
 
 add_subdirectory(${SOC_SERIES})
+zephyr_include_directories(common)
