+++ soc/arm/renesas_rcar/gen4/CMakeLists.txt
@@ -1,0 +1,2 @@
+# Copyright (c) 2023 IoT.bzh
+# SPDX-License-Identifier: Apache-2.0
